bot/requirements.txt
pyTelegramBotAPI
requests
pyjwt